refactor of scheduler and read events
