+++ async/src/main/java/ejarosiewicz/com/async/Thread.kt
@@ -1,0 +1,6 @@
+package ejarosiewicz.com.async
+
+enum class Thread {
+    MAIN,
+    IO
+}
