--- domain/src/main/java/ejarosiewicz/com/eventreminder/domain/async/Thread.kt
@@ -1,6 +1,0 @@
-package ejarosiewicz.com.eventreminder.domain.async
-
-enum class Thread {
-    MAIN,
-    IO
-}
