+++ domain/src/main/java/ejarosiewicz/com/eventreminder/domain/read/DefaultReadEventsUseCase.kt
@@ -1,14 +1,10 @@
 package ejarosiewicz.com.eventreminder.domain.read
 
-//import ejarosiewicz.com.eventreminder.data.EventRepository
-//import ejarosiewicz.com.eventreminder.domain.converter.EventConverter
-//import ejarosiewicz.com.eventreminder.domain.entity.Event
+import ejarosiewicz.com.eventreminder.domain.data.EventRepository
+import ejarosiewicz.com.eventreminder.domain.entity.Event
 
-//class DefaultReadEventsUseCase(private val eventRepository: EventRepository,
-//                               private val eventConverter: EventConverter) : ReadEventsUseCase {
-//
-//    override fun read(): List<Event> = eventRepository
-//            .read()
-//            .map { entity -> eventConverter.convert(entity) }
-//
-//}
+class DefaultReadEventsUseCase(private val eventRepository: EventRepository) : ReadEventsUseCase {
+
+    override fun read(): List<Event> = eventRepository.read()
+
+}
